Refactored the code for system wide install
The default config file is at /etc/libinput-gestures.conf but you can
have a personal configuration file at ~/.config/libinput-gestures.conf.

I added a PNG file for the desktop app.

There is now a user command libinput-gestures-setup to start/stop the
app manually and automatically as a user, or to install/uinstall the app
as root.

These changes are to prepare the app for inclusion in the Arch AUR.