Implement inclined swipe gestures
In addition to the existing left, right, up, and down swipes a user can
now configure inclined swipes: left_up, left_down, right_up, and
right_down if you really want to get tricky. This is somewhat subtle to
physically discriminate in action though so will remain an undocumented
and experimental feature.

If you don't configure any of these inclined swipes, as per the default
configuration, then swipe behaviour remains the same as before this
change, although a small bug was also fixed in the determination of a
minimum threshhold for movement.