Allow libinput-gestures to be run without homedir
This allows libinput-gestures to be run as a root user without a
homedir. Running as root is the most straightforward route to running
ydotool commands from libinput-gestures.

Prior to this changes when libinput-gestures was run as a user without a
home dir relslash.relative_to would raise a TypeError upon receiving
None as an argument.